Separate the driver into its own crate that uses trans, typeck.
