+++ src/librustc_driver/mod.rs
@@ -8,10 +8,3 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#[cfg(rustdoc)]
-extern crate "rustdoc" as this;
-
-#[cfg(rustc)]
-extern crate "rustc_driver" as this;
-
-fn main() { this::main() }
